benchmarking: move glutton's core into internal/benchmarking/glutton
cmd/benchmarking/glutton/main.go held the entire workload: the service,
its otel instruments, the gossip loop, and both transports. None of it
was reachable outside package main, which left internal/benchmarking/glutton
holding nothing but a fake of a server implemented elsewhere.

Move everything except process lifecycle into the internal package, matching
cmd/benchmarking/boomer-glutton. main.go keeps flag parsing, telemetry boot,
the listener, and the shutdown defers. The --mode switch moves out too, since
it selects a transport rather than managing the process, and now returns an
error for an unknown mode instead of calling serverboot.Fatal.

The destination is top-level internal/ because both packages that want to
share glutton's route contract (the fake, and boomer's glutton user class)
sit outside cmd/benchmarking/glutton/.

Renames: gluttonService to Service, newGluttonService to New, newServer to
NewServer, and the mode switch to Handler. sizes.go moves along with its
only caller. No behavior change.
