+++ internal/benchmarking/glutton/glutton.go
@@ -1,0 +1,726 @@
+// Copyright 2026 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// Package glutton implements the benchmarking workload served by
+// cmd/benchmarking/glutton: an API for consuming RAM, disk, and file
+// descriptors, and for gossiping with other glutton instances, over either
+// gRPC or protobuf-over-HTTP. See internal/proto/glutton/glutton.proto.
+package glutton
+
+import (
+	"context"
+	"crypto/rand"
+	"crypto/sha256"
+	"errors"
+	"fmt"
+	"io"
+	"log/slog"
+	"net/http"
+	"os"
+	"path/filepath"
+	"regexp"
+	"strconv"
+	"strings"
+	"sync"
+	"time"
+
+	"github.com/google/uuid"
+	"go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc/otelgrpc"
+	"go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp"
+	"go.opentelemetry.io/otel"
+	"go.opentelemetry.io/otel/attribute"
+	"go.opentelemetry.io/otel/metric"
+	"google.golang.org/grpc"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/credentials/insecure"
+	"google.golang.org/grpc/reflection"
+	"google.golang.org/grpc/status"
+	"google.golang.org/protobuf/proto"
+
+	"github.com/agent-substrate/substrate/internal/ateinterceptors"
+	gluttonpb "github.com/agent-substrate/substrate/internal/proto/glutton"
+)
+
+const meterName = "glutton"
+
+// Handler builds the request handler for the given wire mode. "grpc" serves
+// gRPC alongside the readiness probe on a single listener; "http" serves the
+// protobuf-over-HTTP route table. An unknown mode comes back as an error so
+// the caller decides how to fail.
+func Handler(mode string, svc *Service) (http.Handler, error) {
+	var handler http.Handler
+	switch mode {
+	case "grpc":
+		srv := grpc.NewServer(
+			grpc.StatsHandler(otelgrpc.NewServerHandler()),
+		)
+		gluttonpb.RegisterGluttonServer(srv, svc)
+		reflection.Register(srv)
+		// The readiness probe is an HTTP GET, so gRPC mode serves it next to
+		// the gRPC handler on the same listener.
+		handler = splitGRPC(srv, readyzMux())
+	case "http":
+		// otelhttp at the mux level + per-handler span follows
+		// docs/dev/best-practices/tracing.md: extract incoming context,
+		// then name the span after the operation in each handler.
+		handler = otelhttp.NewHandler(newMux(svc), "/")
+	default:
+		return nil, fmt.Errorf("must be grpc or http: %q", mode)
+	}
+	return handler, nil
+}
+
+// NewServer enables unencrypted HTTP/2 so gRPC works on the plaintext
+// listener, alongside HTTP/1.1 for the readyz probe.
+func NewServer(handler http.Handler) *http.Server {
+	protocols := new(http.Protocols)
+	protocols.SetHTTP1(true)
+	protocols.SetUnencryptedHTTP2(true)
+	return &http.Server{Handler: handler, Protocols: protocols}
+}
+
+// splitGRPC serves gRPC and plain HTTP on one listener: requests with a
+// gRPC content-type go to grpcSrv, everything else to rest. All glutton
+// RPCs are unary, which is what grpc.Server.ServeHTTP supports.
+func splitGRPC(grpcSrv, rest http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		if r.ProtoMajor == 2 && strings.HasPrefix(r.Header.Get("Content-Type"), "application/grpc") {
+			grpcSrv.ServeHTTP(w, r)
+			return
+		}
+		rest.ServeHTTP(w, r)
+	})
+}
+
+// readyzMux serves the readiness probe both modes need: ateom blocks
+// RestoreWorkload until /readyz returns 200, so ResumeActor cannot report
+// success before this listener is reachable.
+func readyzMux() *http.ServeMux {
+	mux := http.NewServeMux()
+	mux.HandleFunc("/readyz", func(w http.ResponseWriter, r *http.Request) {
+		w.WriteHeader(http.StatusOK)
+	})
+	return mux
+}
+
+// newMux builds the HTTP-mode route table on top of the readiness probe.
+func newMux(svc *Service) *http.ServeMux {
+	mux := readyzMux()
+	mux.HandleFunc("/ping", protoRoute("Ping", svc.Ping))
+	mux.HandleFunc("/writedisk", protoRoute("WriteDisk", svc.WriteDisk))
+	mux.HandleFunc("/readdisk", protoRoute("ReadDisk", svc.ReadDisk))
+	mux.HandleFunc("/writeram", protoRoute("WriteRAM", svc.WriteRAM))
+	mux.HandleFunc("/readram", protoRoute("ReadRAM", svc.ReadRAM))
+	return mux
+}
+
+// protoRoute wraps a protobuf handler with POST-only routing, protobuf
+// unmarshaling, status code mapping, and server-timing headers.
+func protoRoute[Req any, Resp proto.Message, PtrReq interface {
+	*Req
+	proto.Message
+}](spanName string, handler func(context.Context, PtrReq) (Resp, error)) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		start := time.Now()
+		if r.Method != http.MethodPost {
+			http.Error(w, "method not allowed", http.StatusMethodNotAllowed)
+			return
+		}
+		body, err := io.ReadAll(r.Body)
+		if err != nil {
+			http.Error(w, err.Error(), http.StatusBadRequest)
+			return
+		}
+		var req Req
+		ptrReq := PtrReq(&req)
+		if err := proto.Unmarshal(body, ptrReq); err != nil {
+			http.Error(w, "unmarshal: "+err.Error(), http.StatusBadRequest)
+			return
+		}
+		ctx, span := otel.Tracer("glutton").Start(r.Context(), spanName)
+		defer span.End()
+		resp, err := handler(ctx, ptrReq)
+		if err != nil {
+			if st, ok := status.FromError(err); ok {
+				switch st.Code() {
+				case codes.InvalidArgument:
+					http.Error(w, st.Message(), http.StatusBadRequest)
+				case codes.NotFound:
+					http.Error(w, st.Message(), http.StatusNotFound)
+				default:
+					http.Error(w, st.Message(), http.StatusInternalServerError)
+				}
+			} else {
+				http.Error(w, err.Error(), http.StatusInternalServerError)
+			}
+			return
+		}
+		out, err := proto.Marshal(resp)
+		if err != nil {
+			http.Error(w, err.Error(), http.StatusInternalServerError)
+			return
+		}
+		// Glutton does not run ateinterceptors, so without this the serve path has no
+		// server-side timing at all. Mirrors the control-plane gRPC trailer so boomer's
+		// elapsedFromMD logic (source=server) works identically over HTTP.
+		w.Header().Set(ateinterceptors.ServerElapsedTrailer,
+			strconv.FormatInt(time.Since(start).Microseconds(), 10))
+		w.Header().Set("Content-Type", "application/x-protobuf")
+		_, _ = w.Write(out)
+	}
+}
+
+// diskKeyRE rejects anything that could escape the data dir or hit a
+// hidden file: only alphanumerics, underscore, and dash are permitted.
+var diskKeyRE = regexp.MustCompile(`^[a-zA-Z0-9_-]+$`)
+
+type Service struct {
+	gluttonpb.UnimplementedGluttonServer
+
+	dataDir string
+
+	// TODO: split this into per-resource locks (ram, fds, peers). A single
+	// global mutex serializes unrelated operations across all three.
+	mu  sync.Mutex
+	ram map[string][]byte
+	// ramCursor is each array's next WRITE_MODE_OVERWRITE_ROTATE offset.
+	// Absent means 0; invalidated whenever the array is reallocated.
+	ramCursor map[string]int
+	fds       []*os.File
+	peers     map[string]*peerGossip
+
+	ramWriteBytes  metric.Int64Counter
+	ramReadBytes   metric.Int64Counter
+	diskWriteBytes metric.Int64Counter
+	diskReadBytes  metric.Int64Counter
+	pingsReceived  metric.Int64Counter
+	gossipSent     metric.Int64Counter
+	gossipLatency  metric.Float64Histogram
+}
+
+type peerGossip struct {
+	host    string
+	delayMs int32
+	cancel  context.CancelFunc
+	done    chan struct{}
+}
+
+// New constructs a Service storing WriteDisk files under dir and registers its
+// otel instruments. The caller is responsible for creating dir and for calling
+// Close to stop any running gossip goroutines.
+func New(dir string) (*Service, error) {
+	s := &Service{
+		dataDir:   dir,
+		ram:       make(map[string][]byte),
+		ramCursor: make(map[string]int),
+		peers:     make(map[string]*peerGossip),
+	}
+
+	m := otel.Meter(meterName)
+
+	var err error
+	s.ramWriteBytes, err = m.Int64Counter(
+		"glutton.ram.write.bytes",
+		metric.WithUnit("By"),
+		metric.WithDescription("Total bytes written to RAM via WriteRAM over the process lifetime."),
+	)
+	if err != nil {
+		return nil, fmt.Errorf("create glutton.ram.write.bytes counter: %w", err)
+	}
+	s.ramReadBytes, err = m.Int64Counter(
+		"glutton.ram.read.bytes",
+		metric.WithUnit("By"),
+		metric.WithDescription("Total bytes walked by ReadRAM over the process lifetime."),
+	)
+	if err != nil {
+		return nil, fmt.Errorf("create glutton.ram.read.bytes counter: %w", err)
+	}
+	s.diskWriteBytes, err = m.Int64Counter(
+		"glutton.disk.write.bytes",
+		metric.WithUnit("By"),
+		metric.WithDescription("Total bytes written to disk via WriteDisk over the process lifetime."),
+	)
+	if err != nil {
+		return nil, fmt.Errorf("create glutton.disk.write.bytes counter: %w", err)
+	}
+	s.diskReadBytes, err = m.Int64Counter(
+		"glutton.disk.read.bytes",
+		metric.WithUnit("By"),
+		metric.WithDescription("Total bytes read from disk via ReadDisk over the process lifetime."),
+	)
+	if err != nil {
+		return nil, fmt.Errorf("create glutton.disk.read.bytes counter: %w", err)
+	}
+	s.pingsReceived, err = m.Int64Counter(
+		"glutton.ping.requests",
+		metric.WithDescription("Number of Ping requests received."),
+	)
+	if err != nil {
+		return nil, fmt.Errorf("create glutton.ping.requests counter: %w", err)
+	}
+	s.gossipSent, err = m.Int64Counter(
+		"glutton.gossip.requests.sent",
+		metric.WithDescription("Number of gossip Ping requests sent per peer."),
+	)
+	if err != nil {
+		return nil, fmt.Errorf("create glutton.gossip.requests.sent counter: %w", err)
+	}
+	s.gossipLatency, err = m.Float64Histogram(
+		"glutton.gossip.latency",
+		metric.WithUnit("s"),
+		metric.WithDescription("Latency of gossip Ping requests per peer."),
+		metric.WithExplicitBucketBoundaries(
+			0.001, 0.0025, 0.005, 0.01, 0.025, 0.05, 0.1, 0.25, 0.5, 1, 2.5, 5, 10,
+		),
+	)
+	if err != nil {
+		return nil, fmt.Errorf("create glutton.gossip.latency histogram: %w", err)
+	}
+
+	fdsOpen, err := m.Int64ObservableGauge(
+		"glutton.fds.open",
+		metric.WithDescription("File descriptors currently held open by OpenFD."),
+	)
+	if err != nil {
+		return nil, fmt.Errorf("create glutton.fds.open gauge: %w", err)
+	}
+	peerDelay, err := m.Int64ObservableGauge(
+		"glutton.gossip.delay",
+		metric.WithUnit("ms"),
+		metric.WithDescription("Configured gossip delay per peer."),
+	)
+	if err != nil {
+		return nil, fmt.Errorf("create glutton.gossip.delay gauge: %w", err)
+	}
+
+	if _, err := m.RegisterCallback(func(_ context.Context, o metric.Observer) error {
+		s.mu.Lock()
+		defer s.mu.Unlock()
+		o.ObserveInt64(fdsOpen, int64(len(s.fds)))
+		for host, p := range s.peers {
+			o.ObserveInt64(peerDelay, int64(p.delayMs), metric.WithAttributes(attribute.String("host", host)))
+		}
+		return nil
+	}, fdsOpen, peerDelay); err != nil {
+		return nil, fmt.Errorf("register glutton observable callback: %w", err)
+	}
+
+	return s, nil
+}
+
+// Close cancels every running gossip goroutine and waits for them to exit.
+func (s *Service) Close() {
+	s.mu.Lock()
+	peers := s.peers
+	s.peers = make(map[string]*peerGossip)
+	s.mu.Unlock()
+	for _, p := range peers {
+		p.cancel()
+		<-p.done
+	}
+}
+
+// Write to RAM, either overwriting previously-used RAM or allocating additional RAM
+// per request instructions. Data written will be random bytes.
+func (s *Service) WriteRAM(ctx context.Context, req *gluttonpb.WriteRAMRequest) (*gluttonpb.WriteRAMResponse, error) {
+	if req.GetKey() == "" {
+		return nil, status.Error(codes.InvalidArgument, "key is required")
+	}
+	sizeBytes, err := parseBytes(req.GetSize())
+	if err != nil {
+		return nil, status.Errorf(codes.InvalidArgument, "size: %v", err)
+	}
+	if sizeBytes < 0 {
+		return nil, status.Error(codes.InvalidArgument, "size must be non-negative")
+	}
+	size := int(sizeBytes)
+
+	switch req.GetWriteMode() {
+	case gluttonpb.WriteMode_WRITE_MODE_TRUNCATE:
+		buf, err := randomBytes(size)
+		if err != nil {
+			return nil, status.Errorf(codes.Internal, "generate random bytes: %v", err)
+		}
+		s.mu.Lock()
+		s.ram[req.GetKey()] = buf
+		delete(s.ramCursor, req.GetKey())
+		s.mu.Unlock()
+	case gluttonpb.WriteMode_WRITE_MODE_OVERWRITE:
+		s.mu.Lock()
+		existing := s.ram[req.GetKey()]
+		if size > len(existing) {
+			existing = make([]byte, size)
+			s.ram[req.GetKey()] = existing
+			delete(s.ramCursor, req.GetKey())
+		}
+		if _, err := rand.Read(existing[:size]); err != nil {
+			s.mu.Unlock()
+			return nil, status.Errorf(codes.Internal, "generate random bytes: %v", err)
+		}
+		s.mu.Unlock()
+	case gluttonpb.WriteMode_WRITE_MODE_OVERWRITE_ROTATE:
+		if err := s.rotateRAM(req.GetKey(), size); err != nil {
+			return nil, err
+		}
+	default:
+		return nil, status.Errorf(codes.InvalidArgument, "unknown write_mode %v", req.GetWriteMode())
+	}
+
+	s.ramWriteBytes.Add(ctx, int64(size))
+	return &gluttonpb.WriteRAMResponse{}, nil
+}
+
+// rotateRAM re-randomizes size bytes starting at the key's cursor, wrapping
+// at the end of the array, then advances the cursor past the write. Repeated
+// rotates therefore walk the whole array instead of re-dirtying the same
+// prefix. The cursor lives in process memory, so it rides along in snapshots
+// and the walk keeps advancing across suspend/resume cycles.
+func (s *Service) rotateRAM(key string, size int) error {
+	s.mu.Lock()
+	defer s.mu.Unlock()
+	existing := s.ram[key]
+	if len(existing) == 0 {
+		return status.Errorf(codes.NotFound, "rotate needs an existing array %q; fill with TRUNCATE first", key)
+	}
+	if size > len(existing) {
+		size = len(existing)
+	}
+	start := s.ramCursor[key]
+	head := existing[start:min(start+size, len(existing))]
+	if _, err := rand.Read(head); err != nil {
+		return status.Errorf(codes.Internal, "generate random bytes: %v", err)
+	}
+	if wrapped := size - len(head); wrapped > 0 {
+		if _, err := rand.Read(existing[:wrapped]); err != nil {
+			return status.Errorf(codes.Internal, "generate random bytes: %v", err)
+		}
+	}
+	s.ramCursor[key] = (start + size) % len(existing)
+	return nil
+}
+
+// pageSize is the stride of the ReadRAM walk: one byte per 4KiB page is
+// enough to force every page resident without the cost of reading them all.
+const pageSize = 4096
+
+// Walk RAM previously written by WriteRAM, reading one byte per page so
+// every touched page must be resident before the response returns. After a
+// demand-paged restore this converts restore-time laziness into measurable
+// read latency.
+func (s *Service) ReadRAM(ctx context.Context, req *gluttonpb.ReadRAMRequest) (*gluttonpb.ReadRAMResponse, error) {
+	if req.GetKey() == "" {
+		return nil, status.Error(codes.InvalidArgument, "key is required")
+	}
+	s.mu.Lock()
+	defer s.mu.Unlock()
+	arr, ok := s.ram[req.GetKey()]
+	if !ok {
+		return nil, status.Errorf(codes.NotFound, "no RAM array %q", req.GetKey())
+	}
+	walk := int64(len(arr))
+	if req.GetSize() != "" {
+		n, err := parseBytes(req.GetSize())
+		if err != nil {
+			return nil, status.Errorf(codes.InvalidArgument, "size: %v", err)
+		}
+		if n < 0 {
+			return nil, status.Error(codes.InvalidArgument, "size must be non-negative")
+		}
+		walk = min(n, walk)
+	}
+	var sum uint32
+	for i := int64(0); i < walk; i += pageSize {
+		sum ^= uint32(arr[i])
+	}
+	s.ramReadBytes.Add(ctx, walk)
+	return &gluttonpb.ReadRAMResponse{Size: walk, Checksum: sum}, nil
+}
+
+// Write to disk using the specified mode. Data written will be random bytes.
+func (s *Service) WriteDisk(ctx context.Context, req *gluttonpb.WriteDiskRequest) (*gluttonpb.WriteDiskResponse, error) {
+	if !diskKeyRE.MatchString(req.GetKey()) {
+		return nil, status.Errorf(codes.InvalidArgument, "key %q must match %s", req.GetKey(), diskKeyRE)
+	}
+	if req.GetSize() < 0 {
+		return nil, status.Error(codes.InvalidArgument, "size must be non-negative")
+	}
+
+	path := filepath.Join(s.dataDir, req.GetKey())
+
+	var flag int
+	switch req.GetWriteMode() {
+	case gluttonpb.WriteMode_WRITE_MODE_TRUNCATE:
+		flag = os.O_RDWR | os.O_CREATE | os.O_TRUNC
+	case gluttonpb.WriteMode_WRITE_MODE_OVERWRITE:
+		// No O_TRUNC: writes go from offset 0 but any bytes beyond size remain.
+		flag = os.O_RDWR | os.O_CREATE
+	default:
+		return nil, status.Errorf(codes.InvalidArgument, "unknown write_mode %v", req.GetWriteMode())
+	}
+
+	f, err := os.OpenFile(path, flag, 0o600)
+	if err != nil {
+		return nil, status.Errorf(codes.Internal, "open %s: %v", path, err)
+	}
+	defer f.Close()
+
+	h := sha256.New()
+	size := int64(req.GetSize())
+	if err := streamRandomBytes(io.MultiWriter(f, h), size); err != nil {
+		return nil, status.Errorf(codes.Internal, "write %s: %v", path, err)
+	}
+
+	// OVERWRITE has no O_TRUNC, bytes from a larger, earlier write will persist.
+	// The cursor is already at size, so folding the remainder into the
+	// same digest completes it without re-reading the prefix.
+	if req.GetWriteMode() == gluttonpb.WriteMode_WRITE_MODE_OVERWRITE {
+		tail, err := io.Copy(h, f)
+		if err != nil {
+			return nil, status.Errorf(codes.Internal, "hash tail %s: %v", path, err)
+		}
+		size += tail
+	}
+
+	s.diskWriteBytes.Add(ctx, int64(req.GetSize()))
+	return &gluttonpb.WriteDiskResponse{Size: size, Sha256: h.Sum(nil)}, nil
+}
+
+func (s *Service) ReadDisk(ctx context.Context, req *gluttonpb.ReadDiskRequest) (*gluttonpb.ReadDiskResponse, error) {
+	if !diskKeyRE.MatchString(req.GetKey()) {
+		return nil, status.Errorf(codes.InvalidArgument, "key %q must match %s", req.GetKey(), diskKeyRE)
+	}
+
+	path := filepath.Join(s.dataDir, req.GetKey())
+
+	f, err := os.Open(path)
+	if err != nil {
+		if errors.Is(err, os.ErrNotExist) {
+			return nil, status.Errorf(codes.NotFound, "file %q not found", req.GetKey())
+		}
+		return nil, status.Errorf(codes.Internal, "open %s: %v", path, err)
+	}
+	defer f.Close()
+
+	h := sha256.New()
+
+	if req.GetReadMode() == gluttonpb.ReadMode_READ_MODE_DIGEST_ONLY {
+		n, err := io.Copy(h, f)
+		if err != nil {
+			return nil, status.Errorf(codes.Internal, "read %s: %v", path, err)
+		}
+		s.diskReadBytes.Add(ctx, n)
+		return &gluttonpb.ReadDiskResponse{
+			Size:   n,
+			Sha256: h.Sum(nil),
+		}, nil
+	}
+
+	data, err := io.ReadAll(io.TeeReader(f, h))
+	if err != nil {
+		return nil, status.Errorf(codes.Internal, "read %s: %v", path, err)
+	}
+
+	s.diskReadBytes.Add(ctx, int64(len(data)))
+	return &gluttonpb.ReadDiskResponse{
+		Size:   int64(len(data)),
+		Sha256: h.Sum(nil),
+		Data:   data,
+	}, nil
+}
+
+// Make sure it has the specified number of file descriptors open. It will open or
+// close file descriptors to hit the desired count (note this count is in addition to the other
+// FDs needed to run the process).
+func (s *Service) OpenFD(_ context.Context, req *gluttonpb.OpenFDRequest) (*gluttonpb.OpenFDResponse, error) {
+	if req.GetCount() < 0 {
+		return nil, status.Error(codes.InvalidArgument, "count must be non-negative")
+	}
+	target := int(req.GetCount())
+
+	s.mu.Lock()
+	defer s.mu.Unlock()
+
+	for len(s.fds) > target {
+		last := len(s.fds) - 1
+		if err := s.fds[last].Close(); err != nil {
+			slog.Warn("Failed to close glutton fd", slog.Any("err", err))
+		}
+		s.fds[last] = nil
+		s.fds = s.fds[:last]
+	}
+	for len(s.fds) < target {
+		f, err := os.Open(os.DevNull)
+		if err != nil {
+			return nil, status.Errorf(codes.Internal, "open %s: %v", os.DevNull, err)
+		}
+		s.fds = append(s.fds, f)
+	}
+	return &gluttonpb.OpenFDResponse{}, nil
+}
+
+// Receive a ping request, echoing the same response back.
+func (s *Service) Ping(ctx context.Context, req *gluttonpb.PingRequest) (*gluttonpb.PingResponse, error) {
+	s.pingsReceived.Add(ctx, 1)
+	return &gluttonpb.PingResponse{Message: req.GetMessage()}, nil
+}
+
+// Sends network traffic to a peer glutton. Messages will be sent
+// on regular intervals separated by delay_ms.
+func (s *Service) Gossip(_ context.Context, req *gluttonpb.GossipRequest) (*gluttonpb.GossipResponse, error) {
+	want := make(map[string]*gluttonpb.Peer, len(req.GetPeers()))
+	for _, p := range req.GetPeers() {
+		if p.GetHost() == "" {
+			return nil, status.Error(codes.InvalidArgument, "peer host is required")
+		}
+		if p.GetDelayMs() <= 0 {
+			return nil, status.Errorf(codes.InvalidArgument, "peer %q delay_ms must be positive", p.GetHost())
+		}
+		want[p.GetHost()] = p
+	}
+
+	s.mu.Lock()
+	var toStop []*peerGossip
+	for host, existing := range s.peers {
+		w, ok := want[host]
+		if !ok || w.GetDelayMs() != existing.delayMs {
+			toStop = append(toStop, existing)
+			delete(s.peers, host)
+		}
+	}
+	var toStart []*gluttonpb.Peer
+	for host, w := range want {
+		if _, ok := s.peers[host]; !ok {
+			toStart = append(toStart, w)
+		}
+	}
+	s.mu.Unlock()
+
+	for _, p := range toStop {
+		p.cancel()
+		<-p.done
+	}
+
+	for _, w := range toStart {
+		gctx, cancel := context.WithCancel(context.Background())
+		pg := &peerGossip{
+			host:    w.GetHost(),
+			delayMs: w.GetDelayMs(),
+			cancel:  cancel,
+			done:    make(chan struct{}),
+		}
+		s.mu.Lock()
+		s.peers[w.GetHost()] = pg
+		s.mu.Unlock()
+		go s.runGossip(gctx, pg)
+	}
+
+	return &gluttonpb.GossipResponse{}, nil
+}
+
+func (s *Service) runGossip(ctx context.Context, pg *peerGossip) {
+	defer close(pg.done)
+
+	// grpc.NewClient resolves and connects lazily; the first RPC surfaces
+	// any failure, so the peer doesn't have to be reachable at start time.
+	conn, err := grpc.NewClient(pg.host,
+		grpc.WithTransportCredentials(insecure.NewCredentials()),
+		grpc.WithStatsHandler(otelgrpc.NewClientHandler()),
+	)
+	if err != nil {
+		slog.ErrorContext(ctx, "Failed to dial gossip peer", slog.String("host", pg.host), slog.Any("err", err))
+		return
+	}
+	defer conn.Close()
+	client := gluttonpb.NewGluttonClient(conn)
+
+	hostAttr := attribute.String("host", pg.host)
+	ticker := time.NewTicker(time.Duration(pg.delayMs) * time.Millisecond)
+	defer ticker.Stop()
+
+	for {
+		select {
+		case <-ctx.Done():
+			return
+		case <-ticker.C:
+		}
+		msg := uuid.NewString()
+		start := time.Now()
+		resp, err := client.Ping(ctx, &gluttonpb.PingRequest{Message: msg})
+		latency := time.Since(start).Seconds()
+		outcome := "ok"
+		cancelled := err != nil && errors.Is(ctx.Err(), context.Canceled)
+		switch {
+		case cancelled:
+			outcome = "cancelled"
+		case err != nil:
+			outcome = "error"
+		}
+		attrs := metric.WithAttributes(hostAttr, attribute.String("outcome", outcome))
+		s.gossipSent.Add(ctx, 1, attrs)
+		s.gossipLatency.Record(ctx, latency, attrs)
+		if cancelled {
+			return
+		}
+		if err != nil {
+			slog.WarnContext(ctx, "Gossip ping failed", slog.String("host", pg.host), slog.Any("err", err))
+			continue
+		}
+		if resp.GetMessage() != msg {
+			slog.WarnContext(ctx, "Gossip ping returned unexpected message",
+				slog.String("host", pg.host),
+				slog.String("sent", msg),
+				slog.String("received", resp.GetMessage()),
+			)
+		}
+	}
+}
+
+func randomBytes(n int) ([]byte, error) {
+	buf := make([]byte, n)
+	if _, err := rand.Read(buf); err != nil {
+		return nil, err
+	}
+	return buf, nil
+}
+
+// streamRandomBytesChunk caps per-syscall random fill and write size so a
+// multi-gigabyte WriteDisk doesn't have to materialize in RAM.
+const streamRandomBytesChunk = 1 << 20 // 1 MiB
+
+// streamRandomBytes writes total random bytes to w sequentially, in
+// streamRandomBytesChunk-sized chunks. The caller is responsible for the
+// file's open mode and starting offset; this writes from the current
+// position forward.
+func streamRandomBytes(w io.Writer, total int64) error {
+	if total <= 0 {
+		return nil
+	}
+	buf := make([]byte, streamRandomBytesChunk)
+	var written int64
+	for written < total {
+		chunk := buf
+		if remaining := total - written; remaining < int64(len(chunk)) {
+			chunk = buf[:remaining]
+		}
+		if _, err := rand.Read(chunk); err != nil {
+			return fmt.Errorf("generate random bytes: %w", err)
+		}
+		n, err := w.Write(chunk)
+		if err != nil {
+			return err
+		}
+		written += int64(n)
+	}
+	return nil
+}
